+++ perl/FSM/Pipeline/HDLGenerator.pm
@@ -189,16 +189,23 @@ sub _target_language_constructor_arg ($value) {
 }
 
 sub generate_hdl_from_file ($self, $fsm_file = undef) {
+    my $pipeline = _generation_receiver_arg($self);
     my $generation_source_path = _generation_source_path_arg($fsm_file);
     return with_fsm_debug_state(
-        { debug_level => ($self->{debug_level} // 0) },
+        { debug_level => ($pipeline->{debug_level} // 0) },
         sub {
             return FSM::Pipeline::SourceGenerationOrchestrator->generate_from_file(
-                pipeline => $self,
+                pipeline => $pipeline,
                 fsm_file => $generation_source_path,
             );
         },
     );
+}
+sub _generation_receiver_arg ($value) {
+    confess "FSM::Pipeline::HDLGenerator expects generate_hdl_from_file(...) invocant to be a blessed FSM::Pipeline::HDLGenerator object"
+        unless blessed($value) && $value->isa(__PACKAGE__);
+
+    return $value;
 }
 sub _generation_source_path_arg ($value) {
     confess "FSM::Pipeline::HDLGenerator expects generate_hdl_from_file(...) argument to be a scalar filesystem path to a .fsm source root"